+++ python/sglang/multimodal_gen/runtime/disaggregation/transport/codec.py
@@ -151,73 +151,6 @@ def send_tensors(
     socket.send_multipart(parts, flags=flags, copy=False)
 
 
-def pack_tensors_async(
-    tensor_fields: dict[str, torch.Tensor | list[torch.Tensor] | None],
-    scalar_fields: dict | None = None,
-    stream: torch.cuda.Stream | None = None,
-) -> tuple[bytes, list["TensorWrapper"], torch.cuda.Event | None]:
-    """Pack tensors with async D2H on the given stream.
-
-    Returns (metadata_bytes, wrappers, d2h_event).
-    Caller must wait on d2h_event before sending wrappers.
-    """
-    descriptors = []
-    buffers = []
-
-    ctx = torch.cuda.stream(stream) if stream is not None else None
-    if ctx is not None:
-        ctx.__enter__()
-
-    for field_name, value in tensor_fields.items():
-        if value is None:
-            continue
-        if isinstance(value, torch.Tensor):
-            wrapper = TensorWrapper(value)
-            descriptors.append(
-                TensorDescriptor(
-                    field_name=field_name,
-                    shape=list(value.shape),
-                    dtype=dtype_to_str(value.dtype),
-                )
-            )
-            buffers.append(wrapper)
-        elif isinstance(value, list):
-            for i, t in enumerate(value):
-                if t is None:
-                    continue
-                if not isinstance(t, torch.Tensor):
-                    raise TypeError(
-                        f"Expected Tensor in list for field '{field_name}', "
-                        f"got {type(t)}"
-                    )
-                wrapper = TensorWrapper(t)
-                descriptors.append(
-                    TensorDescriptor(
-                        field_name=field_name,
-                        shape=list(t.shape),
-                        dtype=dtype_to_str(t.dtype),
-                        list_index=i,
-                    )
-                )
-                buffers.append(wrapper)
-
-    d2h_event = None
-    if stream is not None:
-        d2h_event = torch.cuda.Event()
-        d2h_event.record(stream)
-
-    if ctx is not None:
-        ctx.__exit__(None, None, None)
-
-    metadata = {
-        "tensor_descriptors": [d.to_dict() for d in descriptors],
-        "scalar_fields": scalar_fields or {},
-    }
-    metadata_bytes = json.dumps(metadata, separators=(",", ":")).encode("utf-8")
-
-    return metadata_bytes, buffers, d2h_event
-
-
 def unpack_tensors(
     parts: list,
     device: str | torch.device = "cpu",
